+++ graphql_compiler/tests/test_end_to_end.py
@@ -18,14 +18,11 @@ import pytz
 import six
 
 from .. import graphql_to_gremlin, graphql_to_match
+from ..deserialization import deserialize_argument, deserialize_multiple_arguments
 from ..compiler import compile_graphql_to_gremlin, compile_graphql_to_match
 from ..exceptions import GraphQLInvalidArgumentError
 from ..query_formatting import insert_arguments_into_query
-from ..query_formatting.common import (
-    deserialize_argument,
-    deserialize_multiple_arguments,
-    validate_argument_type,
-)
+from ..query_formatting.common import ensure_arguments_are_provided, validate_argument_type
 from ..schema import GraphQLDate, GraphQLDateTime, GraphQLDecimal, GraphQLSchemaFieldType
 from ..schema.schema_info import CommonSchemaInfo
 from ..typedefs import QueryArgumentGraphQLType
@@ -348,6 +345,7 @@ class QueryFormattingTests(unittest.TestCase):
             "amount": 5,
             "birthday": datetime.date(2014, 2, 5),
         }
+        ensure_arguments_are_provided(expected_types, serialized_arguments)
         self.assertEqual(
             expected_deserialization,
             deserialize_multiple_arguments(serialized_arguments, expected_types),
